refactor(repo): move source files to internal/vhsgen layout
Restructure project to follow idiomatic Go CLI layout with all
vhsgen package files under internal/vhsgen/. Updates all import
paths in test files and cmd/ entry point to match.

AI-Generated-By: Opencode (Claude)
