+++ internal/vhsgen/analyser.go
