+++ internal/vhsgen/doc.go
