+++ internal/vhsgen/generator.go
