+++ internal/vhsgen/golden.go
