+++ internal/vhsgen/mapping.go
