+++ internal/vhsgen/mapping_test.go
@@ -4,7 +4,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/baphled/cukes-vhs"
+	"github.com/baphled/cukes-vhs/internal/vhsgen"
 )
 
 var _ = Describe("TranslateStep", func() {
