+++ internal/vhsgen/renderer.go
