+++ internal/vhsgen/suite_test.go
