+++ internal/vhsgen/templates/base.tape.tmpl
