+++ internal/vhsgen/types.go
