+++ internal/vhsgen/validator.go
