+++ internal/vhsgen/validator_test.go
@@ -7,7 +7,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/baphled/cukes-vhs"
+	"github.com/baphled/cukes-vhs/internal/vhsgen"
 )
 
 // writeASCIIFile writes content to a .ascii file at path and returns the path.
